Remove product service
